+++ pruebas/01-reading-list/santizai/package.json
@@ -1,33 +1,54 @@
 {
-  "name": "prueba-01",
-  "type": "module",
-  "version": "0.0.1",
-  "jest": {
-    "testMatch": ["**/*.test.mjs"]
-  },
-  "scripts": {
-    "dev": "astro dev",
-    "start": "astro dev",
-    "build": "astro build",
-    "preview": "astro preview",
-    "astro": "astro",
-    "test": "jest"
-  },
-  "dependencies": {
-    "@astrojs/react": "^2.2.1",
-    "@astrojs/tailwind": "^4.0.0",
-    "@nanostores/react": "^0.7.1",
-    "@types/react": "^18.0.21",
-    "@types/react-dom": "^18.0.6",
-    "astro": "^2.8.5",
-    "nanostores": "^0.9.3",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "tailwindcss": "^3.0.24"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^14.0.0",
-    "jest": "^29.6.1"
-  }
+    "name": "prueba-01",
+    "type": "module",
+    "version": "0.0.1",
+    "jest": {
+        "transform": {
+            "\\.tsx?$": "ts-jest"
+        },
+        "testMatch": [
+            "**/__tests__/**/*.[jt]s?(x)",
+            "**/?(*.)+(spec|test).[jt]s?(x)",
+            "**/__tests__/**/*.mjs"
+        ],
+        "moduleFileExtensions": [
+            "ts",
+            "tsx",
+            "js",
+            "jsx",
+            "json",
+            "node"
+        ],
+        "transformIgnorePatterns": [
+            "/node_modules/(?!(astro)/)"
+        ]
+    },
+    "scripts": {
+        "dev": "astro dev",
+        "start": "astro dev",
+        "build": "astro build",
+        "preview": "astro preview",
+        "astro": "astro",
+        "test": "jest"
+    },
+    "dependencies": {
+        "@astrojs/react": "^2.2.1",
+        "@astrojs/tailwind": "^4.0.0",
+        "@nanostores/react": "^0.7.1",
+        "@types/react": "^18.0.21",
+        "@types/react-dom": "^18.0.6",
+        "astro": "^2.8.5",
+        "nanostores": "^0.9.3",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
+        "tailwindcss": "^3.0.24"
+    },
+    "devDependencies": {
+        "@jest/globals": "^29.6.1",
+        "@testing-library/jest-dom": "^5.17.0",
+        "@testing-library/react": "^14.0.0",
+        "@types/jest": "^29.5.3",
+        "jest": "^29.6.1",
+        "ts-jest": "^29.1.1"
+    }
 }
